+++ inverse_dynamics_model.py
@@ -4,10 +4,10 @@ import cv2
 from gym3.types import DictType
 from gym import spaces
 
-from lib.action_mapping import CameraHierarchicalMapping, IDMActionMapping
-from lib.actions import ActionTransformer
-from lib.policy import InverseActionPolicy
-from lib.torch_util import default_device_type, set_default_torch_device
+from vpt.action_mapping import CameraHierarchicalMapping, IDMActionMapping
+from vpt.actions import ActionTransformer
+from vpt.policy import InverseActionPolicy
+from vpt.torch_util import default_device_type, set_default_torch_device
 from agent import resize_image, AGENT_RESOLUTION
