+++ vpt/misc.py
