+++ vpt/normalize_ewma.py
