+++ vpt/util.py
@@ -4,10 +4,10 @@ import torch as th
 from torch import nn
 from torch.nn import functional as F
 
-import lib.torch_util as tu
-from lib.masked_attention import MaskedAttention
-from lib.minecraft_util import store_args
-from lib.tree_util import tree_map
+import vpt.torch_util as tu
+from vpt.masked_attention import MaskedAttention
+from vpt.minecraft_util import store_args
+from vpt.tree_util import tree_map
 
 
 def get_module_log_keys_recursive(m: nn.Module):
