+++ vpt/xf.py
@@ -8,9 +8,9 @@ import torch as th
 from torch import nn
 from torch.nn import functional as F
 
-from lib import misc, mlp
-from lib import torch_util as tu
-from lib import util
+from vpt import misc, mlp
+from vpt import torch_util as tu
+from vpt import util
 
 SENTINEL = 0.1337
